requirements.txt
pillow>=10.0.0
selectolax>=0.3.0
aiohttp>=3.9.0
